Fix CI test failures: subprocess uses esbuild instead of tsx, fix shell assumption
- subprocess.test.ts: tsx isn't a project dependency and unavailable in CI.
  Bundle the relay with esbuild in beforeAll and spawn with node instead.
- local-pty-provider.test.ts: hasChildProcesses test assumed default shell
  is zsh (macOS), but CI runs bash. Pin SHELL env var in test setup.
- Extract spawnRelay helper to subprocess-test-utils.ts (max-lines limit).
